added documentations and recipeService Testing

+++ src/main/java/com/prepper/prepper/service/SavedRecipesService.java
@@ -16,16 +16,31 @@ public class SavedRecipesService {
     @Autowired
     private RecipesService recipesService;
 
+    /**
+     *  Retrieve a list of liked recipes given userID
+     *
+     * @param userID the userID
+     * @return A list of saved recipes if found, else return an empty list
+     */
     public List<Recipes> getSavedRecipes(Integer userID) {
         List<SavedRecipes> savedRecipes = savedRecipesRepository.findByUserID(userID);
-        List<Recipes> recipes = new ArrayList<Recipes>();
-        for (SavedRecipes savedRecipe : savedRecipes) {
-            Integer recipeId = savedRecipe.getRecipeID();
-            recipes.add(recipesService.getRecipe(recipeId));
+        if (savedRecipes != null){
+            List<Recipes> recipes = new ArrayList<Recipes>();
+            for (SavedRecipes savedRecipe : savedRecipes) {
+                Integer recipeId = savedRecipe.getRecipeID();
+                recipes.add(recipesService.getRecipe(recipeId));
+            }
+            return recipes;
         }
-        return recipes;
+        return new ArrayList<>();
     }
 
+    /**
+     *  Saves a public recipe to like list
+     *
+     * @param savedRecipe recipe to be saved in like list
+     * @return Recipe saved from the database
+     */
     public SavedRecipes saveRecipe(SavedRecipes savedRecipe) {
         Integer recipeID = savedRecipe.getRecipeID();
         System.out.println(recipeID);
@@ -36,26 +51,45 @@ public class SavedRecipesService {
         return savedRecipesRepository.save(savedRecipe);
     }
 
+    /**
+     *  Remove liked recipe given userID and recipeID
+     *
+     * @param userId, recipeId
+     * @return True if recipe removed from like list, else return false
+     */
     public boolean removeSavedRecipe(Integer userId, Integer recipeId) {
         List<SavedRecipes> savedRecipes = savedRecipesRepository.findByUserID(userId);
-        for (SavedRecipes recipe : savedRecipes) {
-            if (recipe.getRecipeID().equals(recipeId)) {
-                Integer id = recipe.getID();
-                savedRecipesRepository.deleteById(id);
-                Recipes update = recipesService.getRecipe(recipeId);
-                update.setSaves(-1);
-                recipesService.updateRecipe(update);
-                return true;
+        if (savedRecipes != null) {
+            for (SavedRecipes recipe : savedRecipes) {
+                if (recipe.getRecipeID().equals(recipeId)) {
+                    Integer id = recipe.getID();
+                    savedRecipesRepository.deleteById(id);
+                    Recipes update = recipesService.getRecipe(recipeId);
+                    update.setSaves(-1);
+                    recipesService.updateRecipe(update);
+                    return true;
+                }
             }
         }
         return false;
-
     }
 
+    /**
+     *  Retrieve a list of saved recipes given userID (helper function)
+     *
+     * @param userID the userID
+     * @return A list of saved recipes
+     */
     public List<SavedRecipes> getSavedRecipesByUser(Integer userID){
         return savedRecipesRepository.findByUserID(userID);
     }
 
+    /**
+     *  Retrieve a list of saved recipes given recipeID (helper function)
+     *
+     * @param recipeID the recipeID
+     * @return A list of saved recipes
+     */
     public List<SavedRecipes> getSavedRecipesByRecipe(Integer recipeID){
         return savedRecipesRepository.findByRecipeID(recipeID);
     }
